refact: ivp indicador

+++ src/indicadores/ivp/ivp.file.respository.js
@@ -1,5 +1,5 @@
 const fs = require('fs')
-const webscraper = require('./../../web-scraper/bancocentral.ws')
+const webscraper = require('../../web-scraper/bancocentral.ws')
 
 const repositoryFile = `${__dirname}/../../data/ivp.json`
 let dataFromFile = []
@@ -14,18 +14,23 @@ const getDataFromFile = async () => {
 
 const createDataFile = async (forceRebuild=false) => {
 
-    if (!forceRebuild && fs.existsSync(repositoryFile)) return
+    if (!forceRebuild || fs.existsSync(repositoryFile)) return
     
     if (fs.existsSync(repositoryFile)) fs.unlinkSync(repositoryFile)
-
-    console.log('Creating Repository File...')
     
-    const dataFromWebscraper = await webscraper.getIVPData()
-    const jsonString = JSON.stringify(dataFromWebscraper)
+    try {
+        console.log('Creating Repository File...')
+        
+        const dataFromWebscraper = await webscraper.getIVPData()
+        const jsonString = JSON.stringify(dataFromWebscraper)
 
-    fs.writeFileSync(repositoryFile, jsonString, 'utf8')
+        fs.writeFileSync(repositoryFile, jsonString, 'utf8')
+        
+        console.log('Repository File created')
 
-    console.log('Repository File created')
+    } catch (err) {
+        console.error('Cannot create IVP Repository File ' . err)
+    }
 }
 
 const getAll = () => {
@@ -39,8 +44,8 @@ const getByDate = async (date) => {
     return (ivp) ? ivp : {}
 } 
 
-const getByDateRange = (dateFrom, dateTo) => {
-    const ivpsData = getDataFromFile()
+const getByDateRange = async (dateFrom, dateTo) => {
+    const ivpsData = await getDataFromFile()
 
     const ivps = ivpsData.filter(ivpData => ivpData.fecha >= dateFrom && ivpData.fecha <= dateTo)
